+++ projects/ngx-form-object/src/lib/decorators/build-relationship-form-object/build-relationship-form-object.decorator.ts
@@ -1,6 +1,6 @@
 import { FormObject } from '../../form-object/form-object';
 
-export function RelationshipFormObjectBuilder(propertyName: string): MethodDecorator {
+export function BuildRelationshipFormObject(propertyName: string): MethodDecorator {
 	return function (target: FormObject, _key: string | symbol, descriptor: PropertyDescriptor) {
 		target.createFormObjectMethods = target.createFormObjectMethods || new Map();
 		target.createFormObjectMethods.set(propertyName, descriptor.value);