+++ projects/ngx-form-object/src/lib/form-object-builder/form-object-builder.ts
@@ -6,13 +6,10 @@ import { FormStore } from '../form-store/form-store';
 import { capitalize } from '../helpers/helpers';
 import { PropertyOptions } from '../interfaces/property-options.interface';
 import { MODEL_BUILD_CONTROL_METHODS } from '../types/model-metadata.type';
+import { CREATE_FORM_OBJECT_METHODS } from '../types/model-metadata.type';
 
 export class FormObjectBuilder {
-	public formBuilder: FormBuilder;
-
-	constructor() {
-		this.formBuilder = new FormBuilder();
-	}
+	public formBuilder: FormBuilder = new FormBuilder();
 
 	public create(formObject: FormObject): FormStore {
 		const formFields = {};
@@ -149,7 +146,14 @@ export class FormObjectBuilder {
 		relationshipModel: any,
 		propertyOptions: PropertyOptions = {}
 	): FormStore {
-		const createFormObjectFunction = formObject[`create${capitalize(relationshipName.toString())}FormObject`];
+		const relationshipNameString: string = relationshipName.toString();
+
+		// Deprecated in favour of create form object decorators
+		let createFormObjectFunction = formObject[`create${capitalize(relationshipNameString)}FormObject`];
+
+		if (formObject[CREATE_FORM_OBJECT_METHODS] && formObject[CREATE_FORM_OBJECT_METHODS].get(relationshipNameString)) {
+			createFormObjectFunction = formObject[CREATE_FORM_OBJECT_METHODS].get(relationshipNameString);
+		}
 
 		if (createFormObjectFunction) {
 			const modelFormObject: FormObject = createFormObjectFunction.call(