+++ projects/ngx-form-object/src/lib/types/model-metadata.type.ts
@@ -3,3 +3,4 @@ export const MODEL_HAS_ONE_PROPERTIES = Symbol('model_has_one_properties');
 export const MODEL_HAS_MANY_PROPERTIES = Symbol('model_has_many_properties');
 
 export const MODEL_BUILD_CONTROL_METHODS = Symbol('build_control_methods');
+export const CREATE_FORM_OBJECT_METHODS = Symbol('create_form_object_methods');