+++ projects/ngx-form-object/src/public-api.ts
@@ -17,6 +17,7 @@ export * from './lib/decorators/attribute/attribute.decorator';
 export * from './lib/decorators/belongs-to/belongs-to.decorator';
 export * from './lib/decorators/build-control/build-control.decorator';
 export * from './lib/decorators/has-many/has-many.decorator';
+export * from './lib/decorators/build-relationship-form-object/build-relationship-form-object.decorator';
 
 export * from './lib/types/model-metadata.type';